[Feat] Make non-custodial into a separate route without x-state.
